Kill async stack overflow checks
Summary: Async stack overflow checks were used to detect unbounded async
recursion. It's too troublesome to maintain this feature and it
currently blocks an attempt to remove active tracking of currently
running async function.

The side effect of this change is that instead of injecting exception at
random place early (possibly in a context that did not expect any
exceptions), we let the request run until it runs out of memory.

Reviewed By: @billf

Differential Revision: D1285431