Disregard presence of init methods for LdClsPropAddrKnown
Summary: I don't think this check is actually necessary; we need to initialize
static props per-request regardless of whether it has any non-scalar
init methods.  My guess is that we only had this check because we had
nothing like `findSProp()` which gave us accessibility and visibility
without actually initializing some static properties.

This, along with deleting the early-success pass in ldClsPropAddr()
where we checked only the class name, was causing us to take the slow
path in many places for static properties on self (and probably in
general).

Now we have an extra bit in Class to fill with all sorts of fun things
(or at least one fun thing).

Reviewed By: @jdelong

Differential Revision: D1286416