Add NativeDataInfo Sweep mechanism
Summary: DestructFunc will be called when an
object goes out of scope during a request and
is responsible for all cleanup.

SweepFunc will be called at the end of a
request during sweep.  Sweep functions should
not try to clean up any smart allocated resources
since they may have been already swept.

In practice for NDI Classes this translates into
a `void sweep()` magic method.

If your NDI class DOES NOT have sweep() on it,
then the destructor will be called for either case.
This is fine if you do not have any sweepable resources.

If your NDI class DOES have a dispose() method,
it will be called during sweep, while the destructor
will be called for in-request cleanup.

The normal pattern for this looks like:

  class MyNativeData {
   public:
    MyNativeData() { /* new MyNativeData() */ }
    MyNativeData& operator=(const MyNativeData& other) { /* clone $mnd */ }
    ~MyNativeData() {
      /* unset($mnd); */
      // Clean up sweepable resources (if needed, generally not)

      // Chain through to clean up non-sweepables
      sweep();
    }

    void sweep() {
      /* sweeping */
      // Clean up non-sweepable resources
      fclose(m_fp);
    }
  private:
    // Gets cleaned up with ~MyNativeData() in-request, but is ignored for sweep
    String m_path;

    // Must be manually cleaned up, even during sweep
    FILE *m_fp;
  };

Reviewed By: @edwinsmith

Differential Revision: D1270365