Macro-generate more of IRTranslator
Summary: This puts us on the path to incrementally moving IRTranslator's
nonregular logic into other layers and getting rid of it.

The preprocessor/template trickery here took a while to get right, and I
doubt that this is the minimal solution. Here's an approximation of how
I got to this solution:

- The OPCODES macro is the only thing that has info about the bytecodes'
  immediate signatures, so it has to be the thing to generate properly
  typed calls to HhbcTranslator methods by unpacking
  NormalizedInstruction.

- Not all bytecodes have corresponding HhbcTranslator methods, so
  whatever calls are generated by the OPCODES macro must tolerate
  calling nonexistent methods. SFINAE is the natural answer.

- There are pairs of unpackName() methods. The first one gets SFINAE'd
  away by the enable_if if HhbcTranslator doesn't have the method. The
  first argument is nullptr; this is to cause the compiler to prefer the
  HhbcTranslator-calling overload during overload resolution, while
  allowing it to fall back via the implicit conversion to void* if the
  first one gets SFINAE'd out.

There's still a fair amount of irregular logic in IRTranslator, which
will mostly need to be manually dealt with to get rid of
NormalizedInstruction.

The one thing I haven't done yet is hook up the passing of ni.immVec.

Reviewed By: @edwinsmith

Differential Revision: D1288985