Fix Autocomplete globals for namespaces
Summary: First, strip the slashes from the start of the results when we print them.

This is actually sufficient to keep things working for those who aren't using namespaces.

For those that are, they could either be trying to complete something in a namespace, but haven't gotten to the \ yet, or they could be trying to compete something that isn't in the namespace.

So for matching, we strip namespace on both the user input and on the class/fun name.

Reviewed By: @jwatzman

Differential Revision: D1287557