Fix printing of types for autocomplete, autocomplete argument tooltips, and type at position after namespace change
Summary: Add a new function for us to call, Typing_print.full_strip_ns.

The implementation of this is a little cheesy. It gets called whenever we actually add something to the buffer, instead of when we're adding something that could be a namespace. Seems like the refactoring to pass something around in there for stripping/or not or duplicating a bunch of code isn't worthwhile, so I did it this way.

Reviewed By: @jwatzman

Differential Revision: D1289853