Fix iteration bug with immutable collections
Summary: Various implementations for iteration used by the interpreter and the JIT
were not consistent with each other about immutable collections and the
m_version field. In rare cases where the IterInit instruction was jitted
but the IterNext instruction was interpreted, it was possible for a bogus
"Collection was modifiedi during iteration" exception to be thrown when in
fact the collection had not been modified.

This diff fixes the issue by making the IterInit instruction always read
the collection's version and store it into the ArrayIter's m_version field,
regardless of whether the collection is mutable or immutable.

Reviewed By: @elgenie

Differential Revision: D1290953