OS X port of Hack
Summary: There are two things that needed to be ported: inotify and libelf. I
would separate them into different diffs, but that would be tricky to test
since both are needed to build.

==FSEvents port of inotify==
The best API I could find for FSEvents works by adding a watch for a file path,
running a run loop, and then executing a callback when something has changed.
Since the run loop is blocking, I spawn a pthread to run the loop. The main
thread and the run loop thread communicate using a couple pipes and some
lockless linked lists.

Why pipes? Well, there is an API to schedule a callback with the runloop when a
FD is ready to be read. Furthermore, the ocaml code needs to be able to select
some other sockets at the same time it is waiting for fsevents.

Why linked lists? Serializing and deserializing the data in the pipes would be
a pain, and this works fine.

==getsectbyname port of the libelf stuff==
For the most part this is straightforward. The one really annoying thing is
that the apple API, `getsectbyname()`, is a little naive in returning memory
addresses. OS X seems to have ASLR on by default, so the addresses returned
cause segfaults (though of course it works in `lldb` since by default debuggers
tend to disable ASLR). My workaround was to sneak a `-no_pie` argument through
to the linker.

Reviewed By: @jwatzman

Differential Revision: D1378021