Fix free of "" in rfc1867PostHandler
Summary: A recent fix for a leak introduced a bogus free of the
literal string "" in some circumstances. Even in normal circumstances,
it caused the char* to be freed while the apc_rfc1867_data still
had a copy of it. As far as I can see, the current handler never
used it again, but it was entitled to.

This rewrites a few things to use std::string, and avoid any issues
with lifetimes/destruction.

Reviewed By: @LiraNuna

Differential Revision: D1292594