Create phis for locals
Summary: When generating a control flow merge, find any locals that
are defined differently on incoming paths and create the necessary
phis for them.

There's nothing terribly tricky here; the hardest parts were making
sure that FrameState has snapshots for the incoming blocks, and then
pulling out the local state.  I chose to make LocalVec public, but I'd
be open to some slightly different API.

I haven't done anything at all for wide shuffles, even though they can
theoretically happen as a result of this diff.  None of the existing
tests seem to trigger it, so we'll probably want to create a test for
it when we add it.

Reviewed By: @edwinsmith

Differential Revision: D1277847