Remove hhbbc's "single unit" mode for now
Summary: At first I had this mode to optimize a single unit at a time
(without assuming it can see the whole program).  I think it will be
easy enough to redo this later if we decide it's actually worth
building and maintaining.  There was at least one issue with traits
that would need to be solved before it will work (see task), but we
can still add back the hooks in the index in the same way to kill all
class and function resolution if we want this.

Reviewed By: @edwinsmith

Differential Revision: D1293756