Fix InvalidDerivation warnings in CodeError.js
Summary: When flattening traits with interfaces, we pulled
all of the trait's interfaces in, even the the class already
implemented them explicitly. We mostly got away with this,
because we didn't flatten the class until after we called
checkDerivation on it. But another class could cause us to
flatten the traits earlier.

This avoids adding the duplicate interfaces, and does the
flattening first.

Reviewed By: @elgenie

Differential Revision: D1295077