Initial (disabled) support for call-context sensitive return type inference
Summary: This is some initial support for a way of inferring return
types from functions without type hints, or which would be generics
functions in hack, etc.  The idea is to recursively invoke the type
inference algorithm at callsites, to possibly get more information
about the callee based on argument types at the call.

Reviewed By: @swtaarrs

Differential Revision: D1363519