Some cleanup to CppCall
Summary: I'm adding a new call type (for array data-style virtuals),
and the boolean predicates and overloaded constructors looked like
they'd start to get unweildy when there's four.  (It probably would
actually be ok if it weren't for "int" implying vtable index, and it
looks like nothing used virtual anymore, but I did this before I
realized that.)

The duplication of purpose with FuncPtr is pretty obvious after this.
The only thing that one supports that this doesn't is the fssa thing
(calls through an SSA that contains a constant function address),
which we eventually need to stop doing anyway (it causes problems if
an optimization pass changes the types of arguments to the
instruction, which used to be possible with memelim and we probably
want to be able to do again).  At that point we can probably collapse
them.

Reviewed By: @swtaarrs

Differential Revision: D1298637