Replace json decoder with json-c library
The license of the current json decoder used in the ext_json library
presents ackaging issues for debian, replace the library with the
MIT licensed json-c library. The integration code here is a port of
the PECL jsonc extension by Remi Collet.

There are two main outstanding issues:

* hhvm has its own json encoding system which is not being replaced,
  but it makes a single call to utf16_to_utf8 from the previous
  library which is still included in this patch.

* There are a few facebook specific changes to the json decoder to
  implement an FB_LOOSE flag allowing json like '{z:1}' to be valid,
  this has not been reimplemented.

There are probably a variety of smaller issues as well, the patch mearly
appears to work in some circumstances and still needs careful review.