Fix issue with UseVoidReturn
Summary: Depending on the order in which things resolved, it was possible to
get a UseVoidReturn error early during type inference, and then later figure
out that the function really did return a value. Particularly from recursive
functions.

The new test which was added to verify that systemlib was CodeError free has
failed at least once due to that. I've not been able to repro with that test,
but if I check the CodeError.js files for every quick test, a few will always
have a bogus error due to a recursive systemlib function.

This moves the reporting of the error to the final analysis stage, so that
everything has been resolved.

Reviewed By: @jdelong

Differential Revision: D1298608