Add a random offset to the Xenon initial timer
Summary: So that when all of the machines reboot at once, we have a random start
time which will translate into smoother logging.
Keep the Period the same, only touch the first timer.
Random number from 1..10000, which keeps it in the range of RAND_MAX as well as
well as gives us a large pool of starting times.

Reviewed By: @markw65

Differential Revision: D1295985