Bring back closure name hashes
Summary: Closure names are required to be unique---right now an
assertion will fail in non-repo mode in hhbc-translator for CreateCl
if you have non-unique functions that are defined in different files
and create closures.  It looks like it can also potentially call the
wrong closure.  In WholeProgram mode, the emitter is trying to fix up
the names to be globally unique, but there is some kind of race that I
can't reproduce causing occasional issues when hhbbc builds www
(non-unique classes marked with AttrUnique).

Reviewed By: @billf

Differential Revision: D1300423