Make handling of code.hot overflows more robust
Summary: Currently, functions marked as hot are emitted in the code.hot TC
section. If this section fills up, then code.main is used
instead. However, the current mechanism assumes that no single
translation will be bigger than a fixed threshold (8KB).  If HHVM is
unlucky and tries to JIT a function bigger than that when there's not
enough space left, then it crashes (see attached task).  Simply
bumping this threshold is undesirable, because it incurs wasted,
unused space in code.hot.

This diff makes the fall-back mechanism for code.hot overflows more
robust. Instead of failing when a code block fills up, the assembler
now throws an exception.  The JIT catches this exception and then
stops using code.hot.

This diff also improves the error messages when a code block fills up,
by including the code-block name in the message.

Reviewed By: @swtaarrs

Differential Revision: D1297793