Add support for JIT'ing pseudomains
Summary: This diff adds two new IR opcodes, LdGbl and StGbl to support accessing
global variables in the pseudomain. A LdGbl is a LdLoc with a mandatory
guard on it. A StGbl is just a StLoc, but we generate it only for
pseudomains.

Work Items:
  * Implement suppport for ARM. Just by glancing at the code, ARM
    doesn't have support for doing loads with type tests.

  * Turn on inlining of callsites in the pseudomain, fix bugs.

  * HhbcTranslator will punt on bytecodes that would execute the load of
    a global and its guard AFTER doing effectful operations. Might be
    able to do something smarter so that we don't have to punt.
    (BindL, PushL, SetOpL, StaticLocInit, StaticLoc, all minstrs)

This functionality is gated on Eval.JitPseudomain and
hhvm.eval.jit_pseudomain, and defaults to on.

Reviewed By: @swtaarrs

Differential Revision: D1236672