Enable collection buffer sharing for Map
Summary: This diff enables collection buffer sharing for Map. To enable buffer
sharing it was necessary to make some adjustments to how iterators work
for Maps.

Naively sprinkling calls to check for an immutable buffer at the beginning
every mutating method made things a bit inefficient, so I spent some time
to refactor some of the core operations such as add(), set(), grow(), and
reserve().

This diff is stacked on top of D1318651.

Reviewed By: @dariorussi

Differential Revision: D1324858