Kill c_WaitableWaitHandle::m_creator
Summary: m_creator is used as a hint when creating async stack traces. It doesn't
work reliably, as eager execution of async functions can skip multiple
frames when the m_creator is recorded.

Let's kill this hack, so that we can simplify logic for obtaining the
currently running wait handle.

Reviewed By: @billf

Differential Revision: D1303396