Re-enable redefining the file:// wrapper.
Summary: I did this in D1287000, but then @markw65 had to revert
a diff this was stacked on top of and it was simpler to
revert mine along with the other one (D1289030).

This brings that behavior back and fixes some long standing
bad behavior (assuming getWrapperFromURI will always
return a valid wrapper).

Reviewed By: @markw65

Differential Revision: D1291636