Reflow types if splitCriticalEdges removes unreachable blocks
Summary: Otherwise we can leave AssertType opcodes around because they
never learn they are asserting redundant types.  Codegen blows up if
you try to assert a type that widens, so we even need to do this for
correctness, not just optimization (it can't handle materializing a
constant for the type register in the dst).

Reviewed By: @swtaarrs

Differential Revision: D1304777