Initial code relocator
Summary: This just gets the basic functionality enabled. We now
generate code to a new location, and then immediately relocate it
to where it should go. This ensures that things break fast if
we start generating non-relocatbale code.

The next step will be to add alignment info, so we can relocate
between arbitrary addresses, and then adding support for shrinking
jmps should be straightforward

Reviewed By: @edwinsmith

Differential Revision: D1351705