Add an API for profiling arbitrary things to use during PGO translations
Summary: Adds a TargetProfile<> utility that lets you record
information during a profiling translation, and query the results
during an optimized translation.  Also adds a new RDS mode called
"Local", which is data that is thread local but not-zero'd on every
request (we may also find a different use for this).

Nothing uses this yet, but there are few more ideas we can try before
giving up, so @ottoni and I were thinking it's worth putting up a diff
for the API on its own.  (I'm going to try to use it to profile for
array kinds today.)

Reviewed By: @ottoni

Differential Revision: D1312058