Tweak MixedArray::findForInsert
Summary: Right now, when inserting a new element into a MixedArray, we try to
insert into the first hash-probe slot that is either a Tombstone or
Empty. This means that we have to separately keep track of the first
free slot.

If deletes aren't too common, the first free slot should already
be the Empty slot, so we can do away with tracking this
slot separately.

Reviewed By: @edwinsmith

Differential Revision: D1310553