Re-enable recurse.php without frame elision
Summary: I haven't been able to repro this test failure locally. I'm reenabling
here with frame elision off, to establish whether that's the real
problem.

I also changed the name of the function because it was conflicting with
HH\fun, a Hack function that converts a string to a "function reference"
(really a string, but specially tracked by Hack). I don't think that
could have been causing crashes, but this confused the hell out of me
during debugging.

Reviewed By: @alexmalyshev

Differential Revision: D1309345