Some cleanup on the IR Call instruction
Summary: I was looking at how hard it would be to leave the things on
the stack instead of popping them (I think it is doable---I will try
in a little bit), but I did this separately in case that change
doesn't work.  This diff moves the compile time constants to extra
data, (finally) hooks it into the type assertion pass, and fixes the
docs a bit (a bit misleading about the meaning of S3, and still
referred to Type::None wrt a broken optimization we deleted a while
ago).

Reviewed By: @edwinsmith

Differential Revision: D1310673