Add specialized bytecode and IR for the silence operator
Summary: Instead of laboriously encoding this thing's behavior in bytecode, add
two bytecodes StartSilence and EndSilence that turn its effects on and
off. I added IR opcodes and runtime helpers to make the resulting
machine code compact.

Initially I thought I was going to avoid the unnamed local and store the
old reporting level on the stack instead, but that doesn't work: the
result of the silenced expression will end up sitting on top of the old
reporting level when it comes time to restore the reporting level.

Reviewed By: @jdelong

Differential Revision: D1315149