Support far calls to native builtins
Summary: If a builtin is loaded from a DSO with HNI, it may not be
in the low 2GB, so we need to generate an indirect call
to it.  This diff also fixes trampolines so they correctly
can call faraway code, and removes the too-many-trampoline
assert.  (emitCall can recover from running out of tramps).

Fixes #2637

Reviewed By: @markw65

Differential Revision: D1318778