Dispatch whole BB to the interpreter after reaching JitMaxTranslations
Summary: When we hit JitMaxTranslations for a SrcKey, it's because the code is
very polymorphic.  Dispatching just one instruction to the interpreter
in this case is likely to cause the SrcKey for the very next bytecode
instruction to also hit JitMaxTranslations, since only one instruction
is advanced and some of the polymorphic inputs may not have been
consumed yet.  And this can repeat for a sequence of bytecode
instructions, blowing the TC with translations that are likely to fail
the guards.  This diff alleviates the problem by dispatching a whole
basic block to the interpreter in the REQ_INTERPRET service request.

Reviewed By: @swtaarrs

Differential Revision: D1316215