precolor NativeCall, CallBuiltin, Shr/l, and TVs
Summary: Add precoloring to XLS.  Each use gets a register-pair hint,
based on the needs of instructions that define it or use it.
Added automatic support for NativeCall and CallBuiltin since
we can see what registers they will need.  But no support
for the (many) instructions that internally use a call.
we could add many cases for them, too, over time.

Changed the reserved XMM registers from xmm0,1 to xmm6,7
so that xmm0/1 can be used for argument/return value hints.
xmm6/7 have the same encoding size but are much less likely
to be used for parameter passing.

Reviewed By: ottoni

Subscribers: trunkagent, hphp-diffslists, ps

Differential Revision: D1321734

Tasks: 3098678