Allow parser to be generated with bison 3
Summary: We use lots of hacks on the parser output sadly, and removing them is pretty hard (the worst being the definiation of the `k_T_WHATEVER` constants by including the `.hpp` file many times).

After putting a few hour into removing those hacks, I just made the regexes stronger and removed many of the bison3 changes.

This is needed for debian packaging.
Closes #2573

Reviewed By: @jdelong

Differential Revision: D1313556