Use declarative class for config options
Summary: This is the first step to supporting INI everywhere. Instead of methods on the HDF object, we need a single place that will parse whatever you give it.

The next step is to thread a parsed INI object into everywhere that these hdfs are, and then use that too.

I renamed the methods on the hdf class so I could catch them all.

Reviewed By: @fredemmott

Differential Revision: D1321384