actually parse an ini file and pass it down
Summary: This will allow ini setting as well as hdf settings to be used.

It should be noted, that this style is way worse. `ini_get` still won't work on these things, but at least the data will be used. For `ini_get` to work, we need to actually bind the memory locatin instead of returning data. I'll look at that next (for the easy cases).

Reviewed By: @fredemmott

Differential Revision: D1321438