add support for filtering in find
Summary: The find module provides and API that gives bucket of a files at a time.
The reason is to allow the parser to start before the list of files is computed.
This diff adds support for filtering (a filter that is applied *before* the bucket
of file is computed).

Reviewed By: @avikchaudhuri

Differential Revision: D1320801