Fix refcount-opts for $this in static functions
Summary: The changes to refcount-opts.cpp are sufficient to fix the crash but I
decided to do a little related cleanup. I properly implemented the null $this
fatal for This and CheckThis and changed preOptimizeLdThis to generate cleaner
code in cases like this. It won't help perf but the resulting code is much less
embarrassing.

Reviewed By: @alexmalyshev

Differential Revision: D1321977