Fix issues with closures and static dispatch to __invoke methods
Summary: We never used to dispatch to closure invoke methods
statically, so the missing isClosureBody here never could be hit.
(The rule in the runtime right now is that if you call a function with
FPushObjMethod and it has AttrStatic, you only turn the $this into a
Class if it's not a closure body.)

Also put the kUninitializedRIP under Eval.HHIRGenerateAsserts---I turn
that flag to false when I want to see the code we'll generate in an
optimized build, and always have to ignore this part.

Reviewed By: @ottoni

Differential Revision: D1323368