Fix logic to decide when to emit profiling translations
Summary: Occasionally, the JIT was producing profiling prologue (Proflogue)
translations that were never retranslated in Optimize mode.  This diff
fixes a couple of issues that were causing this.

In addition, this diff adds an option to tc-print to filter the top
translations by a specific TransKind.  This was useful to identify the
Proflogue translations that were still reachable.  Analysis of these
translations helped determining the root causes of the problem.

Reviewed By: aravind

Differential Revision: D1323276