Fallback to fp->m_func->cls() if fp->m_this was wiped
Summary: Recently, the responsibility to free locals (including $this) was moved
from ext_asio to the async function itself to match behavior of normal
functions.

The problem is that when the onAsyncFunctionFailed hook is called, $this
is no longer available and when we try to collect an async stack trace
from that hook, the class name is no longer available.

Fallback on using fp->m_func->cls(), which provides __CLASS__ instead of
LSB get_called_class(), but it's better than nothing.

Reviewed By: @billf

Differential Revision: D1322634