Replace json decoder with json-c library
Summary: The license of the current json decoder used in the ext_json library
presents ackaging issues for debian, replace the library with the
MIT licensed json-c library. The integration code here is a port of
the PECL jsonc extension by Remi Collet.

There is one main outstanding issues:

* There are a few facebook specific changes to the json decoder to
  implement an FB_LOOSE flag allowing json like '{z:1}' to be valid,
  this has not been reimplemented.

There are probably a variety of smaller issues as well, the patch mearly
appears to work in some circumstances and still needs careful review.
Closes #2612

Reviewed By: @JoelMarcey

Differential Revision: D1316200

Pulled By: @ptarjan