FrameState merging fix for locals
Summary: When merging FrameState for locals, if a local maps to two
different SSATmps on the incoming branches, we currently discard
the local -> SSATmp mapping. This is problematic for ref-counting,
as we no longer have an SSATmp to consume references for the local at a
later syncing point.

This diff fixes it for the simple case when one of the SSATmps
came from a passthrough instruction with the other SSATmp as the
source. The following IR is an example where this situation can occur:

   B4: (preds B0)
      (13) t6:Arr = LdLoc<Arr,0> t5:FramePtr
      (15) t8:Arr<PackedKind> = CheckType<Arr<PackedKind>> t6:Arr -> B5<Likely>
    -> B9

   B5 (fallthrough) -> B17:  Jmp -> B6

   B9: (fallthrough) -> B10 (fall through) -> B13: Jmp -> B12: Jmp -> B8: Jmp -> B6

   B6: (preds B17 B8)
           t7:Cell = phi t19:CellB17<Likely>, t9:CellB8
      (14) t7:Cell = DefLabel
      --- bc 46, spOff 2 (foo)
    46: Print
      (40) t20:StkPtr = SpillStack t1:StkPtr, 0, t7:Cell
      (41) SyncABIRegs t5:FramePtr, t20:StkPtr
      (42) ReqBindJmp<46>

After the CheckType (instruction 15) in B4, local 0 maps to
t8:Arr<PackedKind> on the next branch (B9), and to t6:Arr on the taken
branch (B5). At a later syncing point (SyncABIRegs (instruction 41)),
we exit with an unconsumed reference for local 0, since we don't have
an SSATmp mapping to it.

Reviewed By: @swtaarrs

Differential Revision: D1324268