Added ** and **= operators to HHVM
Summary: This diff adds the `**` and `**=` operators from PHP 5.6.0beta1 to HHVM.
This is done by creating a brand new opcode "Pow" to represent it.

The logic for the opcode is executed using the already implemented method `f_pow` that maps to the PHP method `pow`.
Please note that at the time of writing, the method `pow` is implemented incorrectly for certain edge cases involving non-numeric values. This will be fixed in a separate diff.

Reviewed By: @ptarjan

Differential Revision: D1288749