Fix a bug in resolve_func_fallback when there is no function to resolve
Summary: It was assuming either the fallback or non-fallback name
would resolve, and didn't test for an empty range.  It handled the
edge-case where the name is found in both places, but not the case
where it's found in neither.  (We were hitting this on a use of
"array_map" inside of a namespace, because hhbbc doesn't analyze
systemlib's hhas unit right now, so array_map doesn't resolve.  The
path that resolves it outside of a namespace works correctly.)

Reviewed By: @scannell

Differential Revision: D1324431