Make sure we access main and stubs consistently
Summary: I want to be able to pass two arbitrary CodeBlocks
into genCode, and have everything go there, to enable jitting
into a scratch area, and then relocating to the target space.

This ensures that the target CodeBlocks are not used (via different
aliases, such as mcg->code.stubs()) while genCode is executing,
and fixes the issues I found.

Reviewed By: @bertmaher

Differential Revision: D1325281