hh_client --search. Improve result ranking
Summary: This is what we used to do in the IDE with Monocle results to make them more relevant, but I'm putting them here so anybody using search can benefit.

Ranking rules: (most important to least)
1. exact matches to the search should be at the top
2. case insensitive exact matches to the search should be next.
3. rank in this order:
  -- classes/funs
  -- typedefs
  -- everything else
4. alphabetical sort on key.

I implemented this in a compare function, so all we have to do is sort the results.

Reviewed By: @jwatzman

Differential Revision: D1325784