Make frame elision slightly more conservative
Summary: recurse.php was crashing because we were eliminating the DefInlineFP's
but not the SpillFrame's of the inlined recursive calls. These are
frames written to the stack but not linked into the AR chain. When the
unwinder tried to unwind from the eventual stack overflow fatal, it
would encounter these frames and try to unwind them as TVs.

What this diff does is to block the elision of any DefInlineFP if its
corresponding SpillFrame is used more than twice: once by the
DefInlineFP we're considering, and once by its corresponding
DefInlineSP.

I think this makes us slightly more conservative than we need to be -- I
believe it is valid sometimes to eliminate a DefInlineFP without
eliminating its corresponding SpillFrame, as long as we know there will
be no unwinding within the region where this gap exists.

Reviewed By: @jdelong

Differential Revision: D1326900