Avoid allocating extra args array for 2-argument min() and max()
Summary: This is pretty straightforward. Add a second declared param to each one,
which defaults to null. Instead of consulting the size of the argv
array, consult argc. Fix the generated ext_hhvm stuff to handle the case
of optional params on a varargs function (I guess this has never been
seen before?).

Reviewed By: @scannell

Differential Revision: D1324578