Cleanup PendingFixups
Summary: When emitting prologues, we just wrote the fixups directly.
This makes everything go through the pending fixup list. This should
make it easier to generate the code into a temporary location, and then
move it to its final destination.

While looking at the code, it seemed to be in the wrong place; it was
implemented on FixupMap, but every use of the pending fixups had to do
mcg->fixupMap()->whatever(), and there were accessors which were provided
solely to let MCGenerator manipulate it. So I moved it to MCGenerator,
which also matches the pending catch list.

Reviewed By: jdelong

Subscribers: hphp-diffslists, ps

Differential Revision: D1328671