Fix server-mode tests, add -vEval.JitRequireWriteLease
Summary: Without the RequireWriteLease option, many of the tests are just
skipping the JIT entirely because they can't get the write lease. And a recent
change to how config files are handled was disabling server mode for all tests.

I also changed it to rerun a test in CLI mode if it fails in server mode. I'm
seeing occasional crashes (real bugs that I have yet to debug) that are
truncating the output for concurrently running tests without making the curl
command fail. Re-running the test in CLI mode will eliminate these spurious
failures, at the expense of masking some tests that are truly broken in server
mode. The verbose output indicates when this happens to allow some kind of
visibility into it.

Reviewed By: @markw65

Differential Revision: D1349388