Fix an imprecise case in union_of that could lead to index invariant violations
Summary: Unioning Obj and ?Obj=Foo would produce InitCell, which is a
bummer.  After a fix to WaitH<T> operator==, we infered a situation
better in www in a way that would cause it to infer less a precise
type on a later analysis round, and make an index type grow, which is
illegal.

Reviewed By: drussi

Subscribers: trunkagent, hphp-diffslists, ps

Differential Revision: D1329963