Fix the JSON extension build stuff
The bundled JSON parser is under a license that Debian doesn't like, so we
need to use another parser for Debian to package HHVM. However, the bundled
parser is better and matches PHP5.5 behaviour, so we'd prefer to use that.
Rather than using json-c if it exists, the usage is explicitly opt-in. So
people that have it installed anyway aren't forced to use it.

This commit does the following:

* Moves the cmake logic into an extension-specific config.cmake file, since
  we can do that now.
* Re-orders the checks in FindLibjsonc to look for the `json-c` version
  first. It seems that the `libjson.so` library only exists for backwards
  compatibility
* Exposes a new option, `USE_JSONC`, that causes the json extension to be
  built with `libjson-c`, causing an error if it can't be found.