Allow setting of RuntimeOptions when building a repo
Summary: The Apache extension defaults to off in cli mode,
which means that it doesn't load by default when building a
repo, which means that whole program analysis doesn't know about
any of its functions. If you want it, you need to enable it at
build time - which means passing in RuntimeOptions.

Reviewed By: bertrand

Subscribers: hphp-diffslists, ps

Differential Revision: D1331751