Kill unnecessary t___construct() methods
Summary: The t___construct() methods were required by hphpc (it would not compile
without them). Since these methods were declared as private, they were
unreachable. The root class of all WaitHandles still defines constructor
as private, which should in theory stop invalid attempts to construct
objects. In practice, it doesn't (and didn't) help, as HHVM tries to build
object first, then decide whether constructor can be called and when
not, destroy the object. That doesn't work well with extension classes,
as they often cannot handle weird state in their destructor. This this
does not have any effect on this behavior and is completely orthogonal.

Also, __construct() was kept in StaticWaitHandle and
ExternalThreadEventWaitHandle due to a buggy design of gen-ext-hhvm,
which emits constructors and destructors only if there is at least one
method. Fix of that bug is out of scope of this diff.

Reviewed By: @billf

Differential Revision: D1346810