incremental mode speedup
Summary: We need to constantly add new dependencies in incremental mode.
The problem is that the list of dependencies for a single object can become pretty
large. When that happens, we have to scan the elements in sequence.

I added a new hash table that is used to check if a dependency exists.
It's a 2X speedup in incremental mode.

Reviewed By: eletuchy

Subscribers: hphp-diffslists, ps, jwatzman, ptc

Differential Revision: D1330760