Bypass getWaitHandle in the common case
Summary: Currently, getWaitHandle() is called when hphpc/hhbbc cannot determine that
the object on stack is already a WaitHandle. Most of the time, the object
on stack is already a WaitHandle (in which case it just returns $this).
Optimize for this case by explicitly checking for
InstanceOfD<WaitHandle> in the emitter.

Special case the implementation of InstanceOfD<WaitHandle> to use
IsWaitHandle, and try to trace through JmpZ/JmpNZ when we know the
likely outcome of the InstanceOfD check.

Reviewed By: @jano

Differential Revision: D1334324