VASM part 6: generate vasm for the whole IRUnit
Summary: Finish converting all the remaining HHIR cg functions to vasm.
This required copying and specializing codeGenImpl(), so I added
that to the BackEnd api. The code duplication with ARM is unfortunate
but I expect it to be temporary; we'll be able to share a lot more
code again once ARM is converted.

Reviewed By: @alexmalyshev

Differential Revision: D1460418